# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For JPO ERP",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/warehouse": {
            "post": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Create a new warehouse",
                "operationId": "createWarehouse",
                "requestBody": {
                    "description": "Choose either store_id or convection_id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "store_id": {
                                        "description": "ID of the store (optional)",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": ""
                                    },
                                    "convection_id": {
                                        "description": "ID of the convection (optional)",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": ""
                                    },
                                    "name": {
                                        "type": "string",
                                        "default": "toko1"
                                    },
                                    "address": {
                                        "type": "string",
                                        "default": "toko1 bandung"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "default": "12345677821"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success create new Warehouse",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/warehouses/": {
            "get": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Get all warehouses",
                "description": "Get All Warehouse",
                "operationId": "indexWarehouse",
                "responses": {
                    "200": {
                        "description": "success fetch Warehouses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Warehouses": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Failed to fetch Warehouses.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/warehouse/{id}": {
            "get": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Get warehouse by id.",
                "description": "Get warehouse by id.",
                "operationId": "showWarehouse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the warehouse",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "store_id": {
                                            "type": "string"
                                        },
                                        "convection_id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "address": {
                                            "type": "string"
                                        },
                                        "phone_number": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch warehouse",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/warehouse/u/{id}": {
            "put": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Update warehouse by ID",
                "description": "Update Warehouse by ID.",
                "operationId": "updateWarehouse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the warehouse to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Warehouse Name"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Updated Warehouse Address"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Warehouse updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Warehouse updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "store_id": {
                                                    "type": "string"
                                                },
                                                "convection_id": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "address": {
                                                    "type": "string"
                                                },
                                                "phone_number": {
                                                    "type": "string"
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                },
                                                "updated_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or failed to update warehouse",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/warehouse/ban/{id}": {
            "post": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Ban warehouse",
                "description": "Ban Warehouse.",
                "operationId": "banWarehouse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the warehouse to ban",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized. Token is missing or invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to ban warehouse",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}